test: Use yukicoder 1649 for tests
