+++ examples/data/yukicoder_1649.sample-1.in
