+++ examples/data/yukicoder_1649.sample-2.in
