+++ examples/yukicoder_1649.py
